x86/e820: do not fixup memmap in copy_e820_map
And instead introduce a new helper to mark the low 1MB VGA/ROM region
as reserved. Note this might be a slight change from current functionality
where copy_e820_map would just leave a hole in the [640KB, 1MB) region
if it was found to be reported as RAM in the memory map.

No functional change expected overall.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: Wei Liu <wei.liu2@citrix.com>