x86/shadow: alloc enough pages so initialization doesn't fail
Current code in shadow_enable will allocate a shadow pool of 4MB
regardless of the values of sh_min_allocation or
shadow_min_acceptable_pages, which means that calls to
shadow_alloc_p2m_page can fail even after the check and allocation
done just above.

Fix this by always checking that the pool is big enough so the rest of
the shadow_init function cannot fail due to lack of pages in the
shadow pool. This is relevant to shadow_alloc_p2m_page which requires
a minimum amount of shadow_min_acceptable_pages(d) + 1 in the pool.

This allows booting a guest using shadow and more than 6 vCPUs.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Tim Deegan <tim@xen.org>
Cc: George Dunlap <george.dunlap@eu.citrix.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: Wei Liu <wei.liu2@citrix.com>
---
I think this should be considered for 4.12, or else attempting to
create a shadow guest with more than 6 vCPUs fails.