ioreq: add fields to allow internal ioreq servers
Internal ioreq servers are plain function handlers implemented inside
of the hypervisor. Note that most fields used by current (external)
ioreq servers are not needed for internal ones, and hence have been
placed inside of a struct and packed in an union together with the
only internal specific field, a function pointer to a handler.

This is required in order to have PCI config accesses forwarded to
external ioreq servers or to internal ones (ie: QEMU emulated devices
vs vPCI passthrough), and is the first step in order to allow
unprivileged domains to use vPCI.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Reviewed-by: Paul Durrant <paul.durrant@citrix.com>
---
Changes since v2:
 - Drop the vcpu parameter from the handler.

Changes since v1:
 - Do not add an internal field to the ioreq server struct, whether a
   server is internal or external can already be inferred from the id.
 - Add an extra parameter to the internal handler in order to pass
   user-provided opaque data to the handler.