x86/hvm: always expose x2APIC feature in max HVM cpuid policy
On hardware without x2APIC support Xen emulated local APIC will
provide such mode, and hence the feature should be set in the maximum
HVM cpuid policy.

Not exposing it in the maximum policy results in HVM domains not
getting such feature exposed unless it's also supported by the
underlying hardware.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
This is a regression, but I'm not able to identify the commit that
introduced the bogus behavior.