x86/hvm: allow ASID flush of running vCPUs
Current implementation of hvm_asid_flush_vcpu is not safe to use
unless the target vCPU is either paused or the currently running one
on the pCPU, as it modifies the generation without any locking.

Fix this by adding a lock to protect the contents of hvm_vcpu_asid,
and use it when accessing any of the fields in the struct. This allows
to safely perform ASID flushes without having to pause vCPUs.

This is not a bugfix as no callers that would violate the assumptions
listed in the first paragraph have been found, but a preparatory
change in order to allow remote flushing of HVM vCPUs.

While there also change the return type of hvm_asid_handle_vmenter
from bool_t to bool.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>